+++ src/components/Footer/Footer.js
@@ -1,11 +1,28 @@
 import { Container, Divider, Typography } from '@material-ui/core'
 import GitHubIcon from '@material-ui/icons/GitHub'
 import LinkedInIcon from '@material-ui/icons/LinkedIn'
-import React from 'react'
+import React, { useEffect, useState } from 'react'
+import { ConditionalRender } from '../ConditionalRender/ConditionalRender'
 import { useStyles } from './style'
 
 export const Footer = () => {
   const classes = useStyles()
+
+  const [date, setDate] = useState(null)
+  useEffect(() => {
+    const getRepository = async () => {
+      const response = await fetch(
+        'https://api.github.com/repos/jsexton-portfolio/portfolio-website'
+      )
+
+      const body = await response.json()
+      const date = new Date(body.pushed_at)
+      const formattedDate = new Intl.DateTimeFormat('en-US').format(date)
+      setDate(formattedDate)
+    }
+
+    getRepository()
+  }, [])
 
   return (
     <div className={classes.root}>
@@ -30,7 +47,14 @@ export const Footer = () => {
         </a>
       </Container>
 
-      <Typography>Last Updated: 2020/03/08</Typography>
+      <ConditionalRender
+        condition={date != null}
+        render={() => (
+          <>
+            <Typography>Last Updated: {date}</Typography>
+          </>
+        )}
+      />
       <Typography>justinsexton.dev@gmail.com</Typography>
       <Typography>&copy; 2020 Justin Sexton</Typography>
     </div>